refactor: Move helper functions for text decoding and PDF extraction to classify.py
